requirements.txt
nba_api
qdrant-client
openai
fastapi
uvicorn
streamlit
python-dotenv
requests